Move CLI Signer utils into clap-utils
