+++ cli/src/cluster_query.rs
@@ -1,5 +1,5 @@
 use crate::{
-    cli::{CliCommand, CliCommandInfo, CliConfig, CliError, DefaultSigner, ProcessResult},
+    cli::{CliCommand, CliCommandInfo, CliConfig, CliError, ProcessResult},
     cli_output::*,
     display::{
         format_labeled_address, new_spinner_progress_bar, println_name_value, println_transaction,
@@ -8,7 +8,9 @@ use crate::{
 };
 use clap::{value_t, value_t_or_exit, App, AppSettings, Arg, ArgMatches, SubCommand};
 use console::{style, Emoji};
-use solana_clap_utils::{commitment::commitment_arg, input_parsers::*, input_validators::*};
+use solana_clap_utils::{
+    commitment::commitment_arg, input_parsers::*, input_validators::*, keypair::DefaultSigner,
+};
 use solana_client::{
     pubsub_client::PubsubClient,
     rpc_client::{GetConfirmedSignaturesForAddress2Config, RpcClient},
